Extract artist specific logic to new ViewModel
