+++ composeApp/src/commonMain/kotlin/io/music_assistant/client/ui/compose/item/ItemDetailsViewModel.kt
@@ -23,11 +23,9 @@ import io.music_assistant.client.data.model.client.items.Podcast
 import io.music_assistant.client.data.model.client.items.PodcastEpisode
 import io.music_assistant.client.data.model.client.items.RecommendationFolder
 import io.music_assistant.client.data.model.client.items.Track
-import io.music_assistant.client.data.model.server.ProviderMapping
 import io.music_assistant.client.data.repository.MediaItemRepository
 import io.music_assistant.client.settings.SettingsRepository
 import io.music_assistant.client.ui.compose.common.DataState
-import io.music_assistant.client.ui.compose.common.map
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
@@ -47,12 +45,6 @@ class ItemDetailsViewModel(
         val albumsState: DataState<List<Album>>,
         val playableItemsState: DataState<List<PlayableItem>>,
         val artistsState: DataState<List<Artist>> = DataState.Loading(),
-        /**
-         * An artist's Albums / Tracks tabs are split into Library / All sub-sections;
-         * these back the ARTIST_ALBUMS / ARTIST_TRACKS tabs instead of [albumsState] /
-         * [playableItemsState]. Null for non-artist items.
-         */
-        val artistSections: ArtistSections = ArtistSections(),
         /** Lazily loaded on demand from the artist overflow menu; NoData until then. */
         val similarArtistsState: DataState<List<Artist>> = DataState.NoData(),
         val albumsSortOption: SortOption? = null,
@@ -94,8 +86,6 @@ class ItemDetailsViewModel(
         ),
     )
     val state = _state.asStateFlow()
-
-    private val fetchArtistItemsUseCase = FetchArtistItemsUseCase(mediaItemRepository)
 
     init {
         // Listen to library changes; refresh the open item + its sub-lists.
@@ -156,16 +146,6 @@ class ItemDetailsViewModel(
     private fun loadSubItems(item: AppMediaItem) {
         when (item) {
             is Artist -> {
-                _state.update {
-                    it.copy(
-                        albumsState = DataState.NoData(),
-                        playableItemsState = DataState.NoData(),
-                        artistSections = ArtistSections.loading(),
-                        albumsSortOption = settingsRepository.getSortOption(SubItemContext.ARTIST_ALBUMS),
-                        playableItemsSortOption = settingsRepository.getSortOption(SubItemContext.ARTIST_TRACKS),
-                    )
-                }
-                loadArtistAlbumSections(item)
                 // Prefetch similar artists in the background so the sheet opens warm: the server's
                 // first per-artist lookup (lastfm + matching) is the slow part, so we pay it while
                 // the user browses albums/tracks rather than on the menu tap.
@@ -237,170 +217,6 @@ class ItemDetailsViewModel(
                         playableItemsSortOption = null,
                     )
                 }
-            }
-        }
-    }
-
-    private suspend fun fetchArtistItems(request: Request): List<AppMediaItem> =
-        mediaItemRepository.fetchMediaItems(request).getOrNull() ?: emptyList()
-
-    private fun loadArtistAlbumSections(artist: Artist) {
-        viewModelScope.launch {
-            try {
-                val library = if (artist.isInLibrary) {
-                    fetchArtistItems(Request.Artist.getAlbums(artist.itemId, artist.provider))
-                        .filterIsInstance<Album>()
-                } else {
-                    emptyList()
-                }
-
-                _state.update {
-                    it.copy(
-                        artistSections = it.artistSections.copy(
-                            library = DataState.Data(
-                                Section(
-                                    items = library.take(ARTIST_SECTION_LIMIT),
-                                    itemList = ItemList.ArtistLibrary(artist.itemId),
-                                ),
-                            ),
-                        ),
-                    )
-                }
-            } catch (e: Exception) {
-                Logger.e("Failed to load artist album sections", e)
-                _state.update {
-                    it.copy(
-                        artistSections = it.artistSections.copy(
-                            library = DataState.Error(),
-                        ),
-                    )
-                }
-            }
-        }
-
-        viewModelScope.launch {
-            val list = fetchArtistItemsUseCase.run(
-                artist,
-                Request.Artist::getAlbums,
-                Album::class,
-            )
-
-            if (list != null) {
-                _state.update {
-                    it.copy(
-                        artistSections = it.artistSections.copy(
-                            all = DataState.Data(
-                                Section(
-                                    list.items.take(ARTIST_SECTION_LIMIT),
-                                    providerDomain = list.mapping.providerDomain,
-                                    itemList = ItemList.ArtistAlbums(
-                                        list.mapping.providerInstance,
-                                        list.mapping.itemId,
-                                    ),
-                                ),
-                            ),
-                        ),
-                    )
-                }
-            } else {
-                _state.update {
-                    it.copy(
-                        artistSections = it.artistSections.copy(
-                            all = DataState.Error(),
-                        ),
-                    )
-                }
-            }
-        }
-
-        viewModelScope.launch {
-            val list = fetchArtistItemsUseCase.run(
-                artist,
-                Request.Artist::getTopTracks,
-                Track::class,
-            )
-
-            if (list != null) {
-                _state.update {
-                    it.copy(
-                        artistSections = it.artistSections.copy(
-                            topTracks = DataState.Data(
-                                Section(
-                                    list.items.take(ARTIST_SECTION_LIMIT),
-                                    providerDomain = list.mapping.providerDomain,
-                                    itemList = ItemList.ArtistTopTracks(
-                                        list.mapping.providerInstance,
-                                        list.mapping.itemId,
-                                    ),
-                                ),
-                            ),
-                        ),
-                    )
-                }
-            } else {
-                _state.update {
-                    it.copy(
-                        artistSections = it.artistSections.copy(
-                            topTracks = DataState.Error(),
-                        ),
-                    )
-                }
-            }
-        }
-    }
-
-    fun loadAlbumsForProvider(mapping: ProviderMapping) {
-        viewModelScope.launch {
-            val itemId = mapping.itemId
-            val providerInstance = mapping.providerInstance
-
-            val albums = fetchArtistItems(
-                Request.Artist.getAlbums(
-                    itemId,
-                    providerInstance,
-                ),
-            ).filterIsInstance<Album>()
-
-            _state.update {
-                it.copy(
-                    artistSections = it.artistSections.copy(
-                        all = DataState.Data(
-                            Section(
-                                albums.take(ARTIST_SECTION_LIMIT),
-                                providerDomain = mapping.providerDomain,
-                                itemList = ItemList.ArtistAlbums(providerInstance, itemId),
-                            ),
-                        ),
-                    ),
-                )
-            }
-        }
-    }
-
-    fun loadTopTracksForProvider(mapping: ProviderMapping) {
-        viewModelScope.launch {
-            val itemId = mapping.itemId
-            val providerInstance = mapping.providerInstance
-
-            val tracks = fetchArtistItems(
-                Request.Artist.getTopTracks(
-                    itemId,
-                    providerInstance,
-                ),
-            ).filterIsInstance<Track>()
-
-            _state.update {
-                it.copy(
-                    artistSections = it.artistSections.copy(
-                        topTracks = DataState.Data(
-                            Section(
-                                tracks.take(ARTIST_SECTION_LIMIT),
-                                providerDomain = mapping.providerDomain,
-                                itemList = ItemList.ArtistTopTracks(providerInstance, itemId),
-                            ),
-                        ),
-                    ),
-                )
             }
         }
     }
@@ -687,21 +503,6 @@ class ItemDetailsViewModel(
             }
 
             is Album -> {
-                _state.value.artistSections.let { sections ->
-                    _state.update { s ->
-                        s.copy(
-                            artistSections = sections.copy(
-                                library = sections.library.map {
-                                    it.copy(items = it.items.replacing(changed))
-                                },
-                                all = sections.all.map {
-                                    it.copy(items = it.items.replacing(changed))
-                                },
-                            ),
-                        )
-                    }
-                }
-
                 val albumsData = (_state.value.albumsState as? DataState.Data)?.data ?: return
                 val updated = albumsData.map { if (it.itemId == changed.itemId) changed else it }
                 rawAlbums = rawAlbums.map { if (it.itemId == changed.itemId) changed else it }
@@ -709,20 +510,6 @@ class ItemDetailsViewModel(
             }
 
             is PlayableItem -> {
-                if (changed is Track) {
-                    _state.value.artistSections.let { sections ->
-                        _state.update { s ->
-                            s.copy(
-                                artistSections = sections.copy(
-                                    topTracks = sections.topTracks.map {
-                                        it.copy(items = it.items.replacing(changed))
-                                    },
-                                ),
-                            )
-                        }
-                    }
-                }
-
                 val tracksData =
                     (_state.value.playableItemsState as? DataState.Data)?.data ?: return
                 val updated = tracksData.map { existing ->
@@ -774,24 +561,8 @@ private fun DataState<out List<*>>.hasItems(): Boolean = when (this) {
     else -> false
 }
 
-data class ArtistSections(
-    val library: DataState<Section<Album>> = DataState.Loading(),
-    val all: DataState<Section<Album>> = DataState.Loading(),
-    val topTracks: DataState<Section<Track>> = DataState.Loading(),
-) {
-    companion object {
-        fun loading() =
-            ArtistSections(DataState.Loading(), DataState.Loading(), DataState.Loading())
-
-        fun error() = ArtistSections(DataState.Error(), DataState.Error(), DataState.Error())
-    }
-}
-
 data class Section<T : AppMediaItem>(
     val items: List<T>,
     val itemList: ItemList? = null,
     val providerDomain: String? = null,
 )
-
-private fun <T : AppMediaItem> List<T>.replacing(changed: T): List<T> =
-    map { if (it.itemId == changed.itemId) changed else it }
